+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [v4](https://github.com/gedzeppelin/monitor-window-switcher/compare/v3...v4) (2021-11-10)
+
+### Feature
+
+- **extension:** Add app switcher support with workplace and window filter.
+
+### Refactor
+
+- **repo:** README updated. 
+
 ## [v3](https://github.com/gedzeppelin/monitor-window-switcher/compare/v2...v3) (2021-04-26)
 
 ### Refactor
@@ -6,6 +16,6 @@
 
 ## [v2](https://github.com/gedzeppelin/monitor-window-switcher/compare/v1...v2) (2021-04-25)
 
-### Features
+### Feature
 
 - **prefs:** add preferences dialog.